+++ src/components/Filterbar/Filterbar.jsx
@@ -223,11 +223,9 @@ function Filterbar({
     {/* Policy hubs has its own release notes */
     ! isPolicyHubs &&
       <div className="absolute top-4 text-xs text-purple-800" style={{left: '110px', fontSize: '0.75rem'}}>
-        versie 2024-10-03<br />
-        - Bekijk hub-stats op mobiel<br />
-        - Betere selectie aanbieders<br />
-      </div>
-}
+        versie 2024-10-30<br />
+        - Nieuwe docs: <a href="https://dashboarddeelmobiliteit.nl/docs/Aanbod/Aanbod-kaart.md"><u>Aanbod-kaart</u></a><br />
+      </div>}
   </>
 }