+++ src/components/Filterbar/FilterbarPolicyHubs.tsx
@@ -344,9 +344,8 @@ function FilterbarPolicyHubs({
       </Modal>}
 
       <div className="absolute top-4 text-xs text-purple-800" style={{left: '110px', fontSize: '0.75rem'}}>
-        versie 2024-10-03<br />
-        - Bekijk hub-stats op mobiel<br />
-        - Betere selectie aanbieders<br />
+        versie 2024-10-30<br />
+        - Nieuwe docs: <a href="https://dashboarddeelmobiliteit.nl/docs/Aanbod/Aanbod-kaart.md"><u>Aanbod-kaart</u></a><br />
       </div>
 
     </>